points icon animation
